Add BiDirectional Naming Scheme for Systems

+++ src/main/kotlin/de/wulkanat/new_frontiers/world/dimension/model/GalacticPosition.kt
@@ -1,3 +1,3 @@
 package de.wulkanat.new_frontiers.world.dimension.model
 
-data class GalacticPosition(val x_ly: Int, val y_ly: Int, val z_ly: Int)
+data class GalacticPosition(val x_ly: Double, val y_ly: Double, val z_ly: Double)
